Split EndPointMetaDataReaderTests
